pyproject.toml
[tool.poetry]
name = "ctf"
version = "0.1.0"
description = ""
authors = ["peterxcli <peterxcli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pycryptodome = "^3.19.0"
sympy = "^1.12"
sage = "^0.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
